+++ js/main.js
@@ -1,3 +1,3 @@
 import './map-form.js';
 import './user-form.js';
-
+import './filters.js';
