+++ js/mock.js
@@ -1,6 +1,4 @@
 export const LENGHT_ADS_ARRAY = 10 ;
-export const USER_PICTURE_PATH = 'img/avatars/user';
-export const USER_PICTURE_EXT = '.png';
 
 export const types = [
   'palace',
@@ -8,27 +6,6 @@ export const types = [
   'house',
   'bungalow',
   'hotel',
-];
-
-export const times = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-export const features = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-export const photos = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
 ];
 
 export const RoomsNumber = {
