+++ js/utils/throttle.js
@@ -16,7 +16,7 @@ function throttle (callback, delayBetweenFrames) {
     // вызываем наш колбэк и перезаписываем lastTime
     // временем "последнего кадра"
     if (now - lastTime >= delayBetweenFrames) {
-      callback.apply(this, rest)
+      callback.apply(this, rest);
       lastTime = now;
     }
   };
